Add generated data, fix file structure, remove an existing research pair
